backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
beanie==1.24.0
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
requests==2.31.0
openai==1.3.0